refactor(node): split conversion into convert/ module
Group marshalling into convert/{mod,job}.rs (room for stats/task_config submodules) — mirrors the Python py_queue split. No behavior change.
